requirements.txt
accelerate==1.11.0
aiofiles==25.1.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.2
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-default-group==1.2.4
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.20.0
filetype==1.2.0
flashtext==2.7
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.2
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multitasking==0.0.12
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.10.2
networkx==3.5
nltk==3.9.2
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
openai==2.7.1
opencv-python==4.12.0.88
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
peewee==3.18.3
pi_heif==1.1.1
pikepdf==10.0.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.11
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==5.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.3
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlite-fts4==1.0.3
sqlite-utils==3.38
srsly==2.5.1
streamlit==1.51.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.9
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.1
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
webencodings==0.5.1
websockets==15.0.1
win32_setctime==1.2.0
wrapt==2.0.0
xxhash==3.6.0
yfinance==0.2.66
zstandard==0.25.0
